rendering closer to react model
