+++ src/App.scss
@@ -61,7 +61,7 @@
   text-align:center;
 }
 
-.homepage, .example-outer{
+.homepage, .example-outr{
   display:flex;
 }
 
@@ -72,7 +72,7 @@
   justify-content:center;
 }
 .example-inner{
-  margin: 5% 15%;
+  margin: 3% 15%;
   &.range{
     width:25vw;
   }
